Phase 4 complete: the GLSL codegen backend
A rule with no finite table is now generated rather than refused. The
generator walks the expression arena in order — children precede
parents, so one forward pass suffices — and emits a typed statement per
node, ending in the aether_rule function SPEC §6 names. The count array
is one statically bounded loop, emitted only when a node asks for a
count, so the contract about data-dependent loops holds by construction.

Three places where C++ and GLSL would have disagreed quietly are now
settled in the spec and obeyed by both paths. A zero divisor is
undefined in GLSL and a trap in C++, so division and modulo by zero
yield zero. Integer arithmetic wraps at thirty-two bits, which the
interpreter gets by computing through unsigned rather than inviting
undefined behaviour, and literals are validated to fit. And the result
is clamped to the state range, because no analysis proves an arithmetic
tree stays in it and a cell outside the range would index past the next
generation's count array. The interpreter and the generator are twins in
the way the two hashes are, and the equivalence test is what says so.

The cache is keyed on shape and, for generated rules, the rule's hash,
so a table rule of unchanged shape still never recompiles and mutation
stays an upload. Compiling costs sixty-one milliseconds on a session's
first rule and under two afterwards, against a budget of two hundred and
fifty, and such a rule steps a 1024² grid at about 2,900 generations a
second.

The phase was specified against a three-dimensional Moore rule reading
individual neighbours — two to the twenty-sixth entries per state as a
table — and that rule is now a fixture agreeing with the oracle over a
thousand generations under every boundary. AV-007 is the other half:
Life as a table and as an expression, through both backends and both
paths, identical four ways. LutRule became CompiledRule on the way,
since the type holds an expression and its GLSL as well as a table, and
CLAUDE.md had still been claiming Lua was unstarted. 219 cases under
ctest.

+++ tests/rule/library_test.cpp
@@ -140,7 +140,7 @@ TEST_CASE("named rules are the rules they claim to be", "[library]") {
     CHECK(cyclic.states == 8);
     CHECK(cyclic.kind == Kind::CountedTotalistic);
     CHECK(std::get<Table>(cyclic.transition).entries.size() == 8 * 9);
-    for (uint16_t own = 0; own < 8; ++own) CHECK(cyclic.counted[own].test((own + 1) % 8));
+    for (uint16_t own = 0; own < 8; ++own) CHECK(cyclic.counted[own].test(static_cast<uint16_t>((own + 1) % 8)));
 
     // The fourteen-state cyclic rule needed 2.8M entries as a full count
     // vector and was left out of the library for it (D-016).
